refactors faq for better reuse
